+++ app/src/main/java/app/revanced/integrations/twitter/settings/SettingsStatus.java
@@ -50,6 +50,7 @@ public class SettingsStatus {
 
     public static boolean profileTabCustomisation = false;
     public static boolean timelineTabCustomisation = false;
+    public static boolean exploreTabCustomisation = false;
     public static boolean sideBarCustomisation = false;
     public static boolean navBarCustomisation = false;
     public static boolean inlineBarCustomisation = false;
@@ -63,8 +64,17 @@ public class SettingsStatus {
     public static boolean unshortenlink = false;
     public static boolean deleteFromDb = false;
     public static boolean nativeDownloader = false;
+    public static boolean nativeTranslator = false;
     public static boolean hideNudgeButton = false;
+    public static boolean hideSocialProof = false;
+    public static boolean customPostFontSize = false;
 
+    public static void customPostFontSize() {
+        customPostFontSize = true;
+    }
+    public static void hideSocialProof() {
+        hideSocialProof = true;
+    }
     public static void hideNudgeButton() {
         hideNudgeButton = true;
     }
@@ -78,6 +88,9 @@ public class SettingsStatus {
 
     public static void nativeDownloader() {
         nativeDownloader = true;
+    }
+    public static void nativeTranslator() {
+        nativeTranslator = true;
     }
 
     public static void deleteFromDb() {
@@ -274,6 +287,10 @@ public class SettingsStatus {
         timelineTabCustomisation = true;
     }
 
+    public static void exploreTabCustomisation() {
+        exploreTabCustomisation = true;
+    }
+
     public static void sideBarCustomisation() {
         sideBarCustomisation = true;
     }
@@ -291,11 +308,11 @@ public class SettingsStatus {
     }
 
     public static boolean enableTimelineSection() {
-        return ( hideNudgeButton || disableAutoTimelineScroll || forceTranslate || hidePromoteButton || hideCommunityNote || hideLiveThreads || hideBanner || hideInlineBmk || showPollResultsEnabled || hideImmersivePlayer || enableVidAutoAdvance || enableForceHD);
+        return ( showSensitiveMedia || hideNudgeButton || disableAutoTimelineScroll || forceTranslate || hidePromoteButton || hideCommunityNote || hideLiveThreads || hideBanner || hideInlineBmk || showPollResultsEnabled || hideImmersivePlayer || enableVidAutoAdvance || enableForceHD);
     }
 
     public static boolean enableMiscSection() {
-        return (roundOffNumbers || enableFontMod || hideRecommendedUsers || hideFAB || hideViewCount || customSharingDomainEnabled || hideFABBtns);
+        return ( nativeTranslator || hideSocialProof || roundOffNumbers || enableFontMod || hideRecommendedUsers || hideFAB || hideViewCount || customSharingDomainEnabled || hideFABBtns);
     }
 
     public static boolean enableAdsSection() {
@@ -311,7 +328,7 @@ public class SettingsStatus {
     }
 
     public static boolean enableCustomisationSection() {
-        return (inlineBarCustomisation || navBarCustomisation || sideBarCustomisation || profileTabCustomisation || timelineTabCustomisation || defaultReplySortFilter);
+        return (exploreTabCustomisation || customPostFontSize || inlineBarCustomisation || navBarCustomisation || sideBarCustomisation || profileTabCustomisation || timelineTabCustomisation || defaultReplySortFilter);
     }
 
     public static void load() {