+++ app/src/main/java/app/revanced/integrations/twitter/settings/fragments/SettingsAboutFragment.java
@@ -85,6 +85,7 @@ public class SettingsAboutFragment extends PreferenceFragment implements Prefere
         flags.put(strRes("piko_pref_feature_flags"),SettingsStatus.featureFlagsEnabled);
         flags.put(strRes("piko_pref_customisation_profiletabs"),SettingsStatus.profileTabCustomisation);
         flags.put(strRes("piko_pref_customisation_timelinetabs"),SettingsStatus.timelineTabCustomisation);
+        flags.put(strRes("piko_pref_customisation_exploretabs"),SettingsStatus.exploreTabCustomisation);
         flags.put(strRes("piko_pref_customisation_navbartabs"),SettingsStatus.navBarCustomisation);
         flags.put(strRes("piko_pref_customisation_sidebartabs"),SettingsStatus.sideBarCustomisation);
         flags.put(strRes("piko_pref_disable_auto_timeline_scroll"),SettingsStatus.disableAutoTimelineScroll);
@@ -116,6 +117,9 @@ public class SettingsAboutFragment extends PreferenceFragment implements Prefere
         flags.put(strRes("piko_pref_customisation_reply_sorting"),SettingsStatus.defaultReplySortFilter);
         flags.put(strEnableRes("piko_pref_force_hd"),SettingsStatus.enableForceHD);
         flags.put(strRes("piko_pref_hide_nudge_button"),SettingsStatus.hideNudgeButton);
+        flags.put(strRes("piko_pref_hide_social_proof"),SettingsStatus.hideSocialProof);
+        flags.put(strRes("translate_tweet_show"),SettingsStatus.nativeTranslator);
+        flags.put(strRes("piko_pref_customisation_post_font_size"),SettingsStatus.customPostFontSize);
 
         LegacyTwitterPreferenceCategory patPref = preferenceCategory(strRes("piko_pref_patches"), screen);